+++ src/ExposeErrorCode.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace NineDigit\eKasa\Client;
+
+/**
+ * Kódy chyby vrátený z Expose
+ */
+class ExposeErrorCode {
+    const TUNNEL_NOT_FOUND = 1;
+}