+++ src/Models/ExposeError.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace NineDigit\eKasa\Client\Models;
+
+class ExposeError {
+    public string $error;
+    public ?int $errorCode = null;
+}